Fix content extraction on App Intents
