+++ App/AppIntents/Intents/GetContentTextIntent.swift
@@ -20,42 +20,19 @@ struct GetContentTextIntent: AppIntent {
     }
 
     @MainActor
-    func perform() async throws -> some IntentResult & ReturnsValue<[ContentBlockEntity]> {
+    func perform() async throws -> some IntentResult & ReturnsValue<String> {
         let database = DatabaseManager.shared
         let articleID = article.articleID
 
         guard let storedArticle = try? database.article(byID: articleID) else {
-            return .result(value: [])
+            return .result(value: "")
         }
 
-        if let cached = try? database.cachedArticleContent(for: articleID),
-           !cached.isEmpty {
-            let blocks = ContentBlock.parse(cached).map(ContentBlockEntity.init(block:))
-            return .result(value: blocks)
+        let extractor = ArticleContentExtractor(article: storedArticle)
+        let extracted = await extractor.extract()
+        guard let text = extracted.text, !text.isEmpty else {
+            return .result(value: "")
         }
-
-        if let raw = storedArticle.content, !raw.isEmpty,
-           let text = ArticleExtractor.extractText(
-            fromHTML: raw,
-            baseURL: URL(string: storedArticle.url),
-            excludeTitle: storedArticle.title
-           ),
-           !text.isEmpty {
-            let blocks = ContentBlock.parse(text).map(ContentBlockEntity.init(block:))
-            return .result(value: blocks)
-        }
-
-        if let articleURL = URL(string: storedArticle.url),
-           let text = await ArticleExtractor.extractText(
-            fromURL: articleURL,
-            excludeTitle: storedArticle.title
-           ),
-           !text.isEmpty {
-            try? database.cacheArticleContent(text, for: articleID)
-            let blocks = ContentBlock.parse(text).map(ContentBlockEntity.init(block:))
-            return .result(value: blocks)
-        }
-
-        return .result(value: [])
+        return .result(value: ContentBlock.plainText(from: text))
     }
 }
