--- Feeds/RSS Parser/ParserVersion.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-nonisolated enum ParserVersion {
-    /// Bump this integer whenever ArticleExtractor logic changes
-    /// to invalidate all cached article content on next launch.
-    static let articleExtractor = 18
-}
