+++ Shared/Strings/Localizable.xcstrings
@@ -23,19 +23,8 @@
             "value": "%1$@ %2$@"
           }
         }
-      }
-    },
-    "%@ [%@](https://developer.apple.com/weatherkit/data-source-attribution/)": {
-      "comment": "A footer that provides a link to the Apple WeatherKit data source attribution page. The first argument is the string “Today.WeatherAttribution.Prefix”. The second argument is the string “Today.WeatherAttribution.Link”.",
-      "isCommentAutoGenerated": true,
-      "localizations": {
-        "en": {
-          "stringUnit": {
-            "state": "new",
-            "value": "%1$@ [%2$@](https://developer.apple.com/weatherkit/data-source-attribution/)"
-          }
-        }
-      }
+      },
+      "shouldTranslate": false
     },
     "%lld": {
       "shouldTranslate": false
@@ -236,18 +225,6 @@
     },
     "Feed ID": {
       "shouldTranslate": false
-    },
-    "FeedEdit.ContentOverrides.Author": {
-      "comment": "A label for the author field in the feed content overrides.",
-      "isCommentAutoGenerated": true
-    },
-    "FeedEdit.ContentOverrides.Body": {
-      "comment": "A label for a picker that lets the user select a field to override in a feed.",
-      "isCommentAutoGenerated": true
-    },
-    "FeedEdit.ContentOverrides.Title": {
-      "comment": "A label for a picker that lets the user select a field to override in a feed.",
-      "isCommentAutoGenerated": true
     },
     "Get latest content": {
       "extractionState": "manual",
@@ -1669,10 +1646,6 @@
           }
         }
       }
-    },
-    "TodayWeather.Location.SetPrompt": {
-      "comment": "A prompt to set the location of the weather.",
-      "isCommentAutoGenerated": true
     },
     "katagaki/SakuraRSS": {
       "localizations": {
